feat + fix: added new tests and fixed bug with null startTime
