+++ test/ru/light/managers/task/FileBackedTaskManagerTest.java
@@ -1,6 +1,5 @@
 package ru.light.managers.task;
 
-import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 import ru.light.managers.history.InMemoryHistoryManager;
 import ru.light.task.EpicTask;
@@ -15,37 +14,50 @@ import java.time.LocalDateTime;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-class FileBackedTaskManagerTest {
+class FileBackedTaskManagerTest extends TaskManagerTest<FileBackedTaskManager> {
+    File file;
 
-    static TaskManager taskManager;
-    static File file;
 
-    @BeforeAll
-    static void setUp() throws IOException {
-        file = File.createTempFile("data", "csv");
-        taskManager = new FileBackedTaskManager(new InMemoryHistoryManager(), file);
-
-        Task newTask = new Task("Title", "Task Description", null, TaskStatus.NEW,
-                Duration.ofHours(1),
-                LocalDateTime.of(2020, 1, 1, 1, 1));
-        taskManager.addTask(newTask);
-        EpicTask epic = new EpicTask("Epic Title", "Epic desc", null);
-        taskManager.addTask(epic);
-        SubTask subTask = new SubTask("Sub Title", "Sub desc", null,
-                TaskStatus.DONE, epic.getId(), Duration.ofHours(1),
-                LocalDateTime.of(2020, 1, 1, 1, 1));
-        taskManager.addTask(subTask);
-
-        taskManager.getById(1);
-        taskManager.getById(2);
+    @Override
+    public FileBackedTaskManager createTaskManager() {
+        try {
+            file = File.createTempFile("data", "csv");
+        } catch (IOException e) {
+            throw new RuntimeException(e);
+        }
+        return new FileBackedTaskManager(new InMemoryHistoryManager(), file);
     }
 
     @Test
     void testLoadFromFile() {
+        FileBackedTaskManager fileBackedTaskManager = getFileBackedTaskManager();
+
+        fileBackedTaskManager.getById(1);
+        fileBackedTaskManager.getById(2);
         FileBackedTaskManager taskManagerFromFile = FileBackedTaskManager.loadFromFile(file);
-        assertEquals(taskManager.getAllStandardTasks(), taskManagerFromFile.getAllStandardTasks());
-        assertEquals(taskManager.getAllEpicsTasks(), taskManagerFromFile.getAllEpicsTasks());
-        assertEquals(taskManager.getAllSubTasks(), taskManagerFromFile.getAllSubTasks());
-        assertEquals(taskManager.getHistory(), taskManagerFromFile.getHistory());
+        assertEquals(fileBackedTaskManager.getAllStandardTasks(), taskManagerFromFile.getAllStandardTasks());
+        assertEquals(fileBackedTaskManager.getAllEpicsTasks(), taskManagerFromFile.getAllEpicsTasks());
+        assertEquals(fileBackedTaskManager.getAllSubTasks(), taskManagerFromFile.getAllSubTasks());
+        assertEquals(fileBackedTaskManager.getHistory(), taskManagerFromFile.getHistory());
+    }
+
+    private FileBackedTaskManager getFileBackedTaskManager() {
+        try {
+            file = File.createTempFile("data", "csv");
+        } catch (IOException e) {
+            throw new RuntimeException(e);
+        }
+        FileBackedTaskManager fileBackedTaskManager = new FileBackedTaskManager(new InMemoryHistoryManager(), file);
+        Task newTask = new Task("Title", "Task Description", null, TaskStatus.NEW,
+                Duration.ofHours(1),
+                LocalDateTime.of(2020, 1, 1, 1, 1));
+        fileBackedTaskManager.addTask(newTask);
+        EpicTask epic = new EpicTask("Epic Title", "Epic desc", null);
+        fileBackedTaskManager.addTask(epic);
+        SubTask subTask = new SubTask("Sub Title", "Sub desc", null,
+                TaskStatus.DONE, epic.getId(), Duration.ofHours(1),
+                LocalDateTime.of(2020, 1, 1, 1, 1));
+        fileBackedTaskManager.addTask(subTask);
+        return fileBackedTaskManager;
     }
 }
